fix: move co-located test files to __tests__ to fix tsc build
Test files in src/components/logs/ were being included by tsc despite
the exclude pattern in tsconfig.app.json. Moved them to
src/components/__tests__/logs/ which is properly excluded.
