+++ web/src/components/__tests__/logs/LogsErrorState.test.tsx
@@ -1,7 +1,7 @@
 import { screen } from "@testing-library/react";
 import { describe, expect, it } from "vitest";
-import { renderWithProviders } from "../../test/utils";
-import { LogsErrorState } from "./LogsErrorState";
+import { renderWithProviders } from "../../../test/utils";
+import { LogsErrorState } from "../../logs/LogsErrorState";
 
 describe("LogsErrorState", () => {
 	it("renders the error message", () => {
